Updating References after Move
JavaLogFactory naming and reference updates.
