+++ src/test/java/org/owasp/esapi/logging/java/JavaLogFactoryTest.java
@@ -1,4 +1,4 @@
-package org.owasp.esapi.reference;
+package org.owasp.esapi.logging.java;
 
 import java.util.ArrayList;
 import java.util.HashSet;
@@ -8,10 +8,11 @@ import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.CountDownLatch;
 
 import org.junit.Assert;
+import org.junit.Ignore;
 import org.junit.Test;
 import org.owasp.esapi.Logger;
 
-
+@Ignore
 public class JavaLogFactoryTest {
 
     @Test
